typedefed the ids used by models/meshes/shaders/etc just in case I want to change the size later added expected vertex and model instance data formats to shader changed model instance transform to general instance state in anticipation of text rendering in particular -convenience functions to modify transform assumes a "position" state, which the scene loader DOES insert for now, but is not there by default so beware -it is expected in the future that manipulation of model instances will be done by scene management classes, not directly cleaned up some type conversion warnings

+++ Renderer/Include/Renderer/RenderOptions.h
@@ -8,11 +8,11 @@
 struct RenderOptions
 {
 	std::shared_ptr<FrameBuffer> frameBuffer;
-	glm::vec2 viewportOrigin;
-	glm::vec2 viewportDimensions;
+	glm::ivec2 viewportOrigin;
+	glm::ivec2 viewportDimensions;
 	bool wireframe;
 
 	RenderOptions();
-	RenderOptions(std::shared_ptr<FrameBuffer> frameBuffer, glm::vec2 viewportOrigin, glm::vec2 viewportDimensions, bool wireframe);
+	RenderOptions(std::shared_ptr<FrameBuffer> frameBuffer, glm::ivec2 viewportOrigin, glm::ivec2 viewportDimensions, bool wireframe);
 };
