+++ Renderer/src/Renderer/RenderOptions.cpp
@@ -6,7 +6,7 @@ RenderOptions::RenderOptions() :frameBuffer(std::make_shared<FrameBuffer>()), vi
 {
 }
 
-RenderOptions::RenderOptions(std::shared_ptr<FrameBuffer> frameBuffer, glm::vec2 viewportOrigin, glm::vec2 viewportDimensions, bool wireframe)
+RenderOptions::RenderOptions(std::shared_ptr<FrameBuffer> frameBuffer, glm::ivec2 viewportOrigin, glm::ivec2 viewportDimensions, bool wireframe)
 	:frameBuffer(frameBuffer), viewportOrigin(viewportOrigin), viewportDimensions(viewportDimensions), wireframe(wireframe)
 {
 }
